Handle ‡3 in 382 fields as ‡v
The significance of the ‡3 is slightly different than the ‡v, but we are
already treating the ‡v as though it might be used as a ‡3, due to the
standard previously not allowing a ‡3 in this field. The easiest way to
incorporate the ‡3 is to add it to the handling logic for ‡v, which will
have the reciprocal effect of treating the ‡3 as a ‡v if it is ever used
that way, which may not happen.