Ensure holdings library list order is deterministic
The list of libraries in which a book resides for the purposes of title
linking is non-deterministic in sequence, leading to title link updates
when the order changes. Managing this list in a LinkedHashSet instead of
a HashSet should resolve the problem, ensuring that the list retains
insertion order. Insertion order itself should be deterministic as the
holdings and items lists are managed in TreeMaps, which themselves
ensure iteration in order by id*.

(*Currently, the holdings are order by id alphabetically, and the items
by id numerically, but each is at least deterministic.)

DISCOVERYACCESS-3071