Process ISBN through explicit ResultSetToFields
We want to display the 020 subfields |a and |q, but only for fields that
have a subfield |a. The StandardMarcFieldMaker doesn't handle that kind
of conditional.
DISCOVERYACCESS-2661

At the same time this was a good opportunity to make sure that the ISBN
handling is compatible with the upcoming PCC MARC standard:
DISCOVERYACCESS-3242