Eliminate dependency on Apache's StringUtils
As of Java 8, the String object natively supports a join() method, which
is the only use of StringUtils we have had that can't be accomplished as
easily without it. Unlike StringUtils.join(), the String.join() method
does not call toString() on the elements to be joined, making it
necessary to make sure that the values passed in are already some form
of CharSequence.