Populate URL fields in Solr with link-specific TOU hooks
The new field url_access_json is intended to replace the
url_access_display Solr field. It has four component fields. "url" and
"description" mirror information previously available in the
url_access_display, while the new, optional, "dbcode" and "providercode"
fields will be populated with link-specific TOU hooks where available.
The "description" argument will also be unpopulated when no description
is available for the link.

The root-level Solr fields, "dbcode" and "providercode" will continue to
be populated, but only with the codes that appear in record-level 899
fields, not the link-specific ones that appear in 856 fields.

DISCOVERYACCESS-2947