Take Unicode character out of regex
I'm not sure why the Unicode character here caused the related search fields
to fail to identify matches, but it did. With the new configuration, the
search fields will not be able to distinguish between words at the beginning
of a text field and words in the middle of a text field immediately following
a word starting "bgn". This seems unlikely enough not to make a noticeable
impact on real life searching.