Include SSID in url_access_json
Rather than add explicit support for ssid, I will simplify the logic by
not enforcing the set of allowable fields to be imported into the json.
The first time, enforcing the set of fields caused problems when a new
field was added that we didn't know about. The risk of occassionally
pulling in erroneous fields seems modest and not destructive if it did
occur.

DISCOVERYACCESS-3438