Limit 041 language codes displayed in item view
Per DISCOVERYACCESS-3404, language codes for the language of an original
version of a work do not necessarily represent the language of the item
described. The languages identified for the table of contents and
summary/abstract are still displayed but not included in the facet.
Hopefully this doesn't come up often enough to cause confusion, but in a
future iteration we may consider displaying annotations with the
languages to represent their role in the work.