Streamline population of subject search fields
We don't really need separate fields for subject_t and subject_addl_t,
and the current implementation is resulting in nine extra
StandardMARCFieldMakers being created for every processed bib. Since
these are a major cause of inefficiency in the code base, I hope for a
minor speed boost to come from this change.