Simplify CitationReferenceNote.java
This should not cause any significant change in behavior, but requires a
set of JUnit tests before merging.

One important simplication here is that the display values are not being
separately handled as valueMain and values880. This is something that
was done to ensure that the original script value would display ahead of
the romanized value where applicable. This is no longer necessary, as
the FieldSet class will now iterate fields in the correct display order
by default (by contrast the the MarcRecord itself, which will iterate
the fields in correct RECORD order).

Minor ways in which I expect behavior to change, both related to
incorrect encodings of values present in both an original, non-Roman
form and a Romanized form:

* If the first indicator of a non-Roman script 510 citation note is not
set to a valid value, then the non-Roman field will no longer be
displayed using the relationship suggested by the matching Romanized 510
citation note, (should one exist), but will be dropped from display.

* In a matching pair of Romanized and non-Roman 510 citation notes, if
the indicators on each field disagree about the appropriate displayed
relationship, the one that will take presendence in display will be the
value on the non-Roman field, where the previous code would be
non-deterministic in which would take presedence.