Create Builder for FieldSet object that sorts fields
In the previous implementation, each DataField was subclassed into an
FSDataField for the purpose of having a separate comparator which sorts
fields in display order instead of record order. The new implemenation
sorts the fields into an ArrayList when the FieldSet object is built.

To enforce the use of the FieldSet.Builder to sort the fields, FieldSet
values are now private and can only set from in class. Some direct field
access is still possible from within MarcRecord.java, but any outside
code is forced to use the new instantiation models.