Add ability to segregate 76x-78x fields into author and title
This implementation isn't complete, as the changes haven't been fully
applied to TitleChange.java, but the segregation seems to be working
well, and tests are available for the segregation in
MarcRecordAuthorTitleSegregationTest.java.

DISCOVERYACCESS-3445 (Part 1)

I also added a more streamlined way to create DataField objects for
tests, which at this point I don't have any plans to apply to production
processing.