Resurrect old begins with boost analysis; Add phrase search modes
DISCOVERYACCESS-2020
The previously adjusted begins-with boost has the advantage of
preventing the first word in an unquoted search from triggering a
begins-with boost due to a match on the third word of a title. It does
not prevent the third word in an unquoted search from triggering a
begins-with boost when matching the first word of a title.

The new plan is to apply the begins with boosting against only quoted
searches and the quoted segment of modified unquoted searches. In that
interest, we have phrase search modes for all-fields and title search
that include the old style begins-with boost field. Once the front end
adjustments are made to employ the new search modes, the modified boost
fields can be dropped from the default all-fields and title search
modes.