Don't equate "et" with "and" in text truncation search
DISCOVERYACCESS-3401
The new behavior should continue to treat the ampersand (&) as
equivalent to the word "and", but the conflation of "and" and "et" was
having unexpected results when combined with the behavior of truncation
searching. That is, the search token 'e' should match words starting
with 'e', but due to this equivalence it would match the word 'and',
while the search token 'a' would not.