Update Subject.java; Add SubjectTest.java
The new standard ResultSetUtilities.SolrFields class needed updating to
support boolean fields, which will come in handling when updating future
fields. I also marked the static method
ResultSetUtilities.resultSetsToSetsofMarcFields() method as deprecated,
as it incompatible with the standard method signature for
generateSolrFields( MarcRecord, SolrBuildConfig ) that I'm now trying to
establish as a pattern.