Eliminate project code dependencies in MarcRecord.java
To allow the marc code package to be portable, it should not use code
from the project that is outside the package. To that end, the methods
from MarcRecord that populate Marc fields based on Jena SPARQL query
results are now in their own class outside the package, and the method
that used StandardMARCFieldMaker.VernMode as an argument, now accepts a
boolean instead.