Eliminate project code dependencies in DataField.java
To allow the marc code package to be portable, it should not use code
from the project that is outside the package. To that end, the
FieldValues subclass has been moved out of DataField.java and into its
own utilities class, along with relevant methods of DataField. Also, the
method that strips initial articles from title string values has been
imported to DataField.java out of the CharacterSetUtils class.