Process at least one batch in ProcessQueue
Quitting time for ProcessQueue is determined by the hour, so if a single
batch is processed it should still be quitting time after the fact.
Moving the quitting time confirmation to the bottom of the loop is
intended to make it easier to process single batches by launching
ProcessQueue with the current hour set as quitting time.