Apply milder matching betw. Roman/Non-Roman name pairs
When there are exactly two names in a record, and they have matching
tags, it's reasonable to encourage them to link even if they aren't
encoded to link. Pushing all names to link, though, has the effect of
interfering with a good link between a name pair in the presence of an
extraneous, unlinked name entry.