Drop LocalDataMaker infrastucture
This infrastructure for creating temporary, local triplestores that are
record-specific for targetting SPARQL queries was added speculatively
and never developed to the point of active use. At this point, the
project development is moving in other directions, so there's no need to
leave the proto-distinction between "main" and "local" queries.