Start to implement GenerateSolrFields.java
So far, the code is working with three sample SolrFieldGenerator
classes, AuthorTitle, Subject, and SimpleProc. The bibliographic MARC
record is divided into MARC segments according to each generator's
supported fields. Then, the fields are generated and serialized into
JSON. There are a lot of things left to implement:

* All relavant generators need to be integrated.
* MARC record segments have to be compared to previous segments so
unchanged segments may not be regenerated.
* JSON field segments have to be compared to previous segments so
unchanged Solr documents may not be re-submitted to Solr.
* JSON field segments need their path to Solr completed. Printing to
stdout is a good confirmation during development, but not a good final
destination.
* Holdings data needs to be integrated to support certain generators
* DocumentPostProcessors must either be integrated into the new workflow
or converted to standard generators.
* HoldingsAndItems.java must be either supported by the new model or be
replaced by functionality in the da-solr-availability project first.