Move "Electronic books" facet block to Subject.java
SuppressUnwantedValues.java was set up in 2013 in response to
DISCOVERYACCESS-498 (commit c425c560cd6df872c1ab00a28875777473860f23)
with the idea that we might have other circumstances where we want to
target particular values to be blocked from specific fields. That hasn't
happened, and even if it does come up, the code is more maintainable
when it's managed in a single class or at least a single execution
stack.

Note that only the subject_topic_facet value is blocked, which has the
effect of blocking the value from the facet, but not search, record
display or subject browse. If there weren't an authority record for this
subject, then this _would_ also block the term from subject browse.