Support 020‡ q values separated by colons
The current standard specifies a semicolon at the end of each ‡q if
followed by another ‡q. (e.g. " ‡a 123456789 ‡q (book 3 ; ‡q np)").
There's precedent to using a colon instead of the semicolon, so the
display standardization should watch for that, and display "123456789
(book 3 ; np)" rather than "123456789 (book 3 : ; np)".