Migrate RecordBoost DocPostProc to RecordBoost SolrFieldGen
The old model involves loading identities of records that may be boosted
into a triplestore, querying the triplestore to find the boost status,
putting a temporary boost flag Solr field into the SolrInputDocument,
and then, in a DocumentPostProcess, finding that temporary field,
removing it, and assigning a boost to the SolrInputDocument.

This is ridiculously convoluted, even if I weren't working to eliminate
as many DocumentPostProcess-ors as possible. The new RecordBoost
SolrFieldGenerator keeps a static in-memory list of boosted records, so
it may become unsustainable if we ever want to apply varied boosts at a
large scale. As unlikely as that is to happen, in that case we would
also want to revisit the process to make the identities of boosted
records more dynamic than a flat file that's part of the Git repo.

Still, the list of 50 boosted records hasn't changed in the last five
years, so this is probably enough for now.