Combine individual segment db tables into wide table
Most of the processing time for this class was IO, and the majority of
that was MySQL communication. In order to limit MySQL interactions, the
process now pulls a single wide row from the database for a bib, does
all the processing, and if anything has changed pushes the row back to
MySQL in a single update query. The process is still IO-bound, but MySQL
interactions are a significantly smaller part of that puzzle.