Supply period in cases where it would be not removed
This turned out to be a pretty straight forward adjustment to the first
implementation. When the criteria appear that would cause us to _not_
remove a trailing period, then we want to add that period should it not
already appear. This runs the risk of adding periods where they really
shouldn't be, but in the absense of any known examples where that might
be a problem, it should be a very small risk. On the other hand,
supplying the period where it is clearly missing in the original record
allows the display of these periods in the facets without the risk of
inconsistent cataloging causing facet values to fail to group based on
those periods. (See the vance screenshot on the Jira ticket.)

DISCOVERYACCESS-3703