Reduce redundancy of indexed title data
The main title of a record is the most heavily indexed and re-indexed
piece of metadata in the index (with and without stemming, with
hyphen/apostrophe grouping vs separating, with and without initial
article, exact match, exact match excluding subtitle). The size of the
index can be reduced notably if in those cases where the title does not
start with an article we don't also index the title as without initial
article. This may require some rebalancing of search weights, and we may
want to go back to indexing these titles twice in title_t while making
the left-anchored search boost copy fields into copy fields of the
title_exact field instead.