Update acquired date fields for date range faceting
DISCOVERYACCESS-3884

The "acquired" integer field is being replaced with the "acquired_dt"
datetime field. As a side effect, the "acquired_month" string field is
switching from YYYYMM format to YYYY-MM format. The second was not a
requested change, but because the "acquired_month" field is not
currently employed by any production services, and because this seems
like a format improvement, I'm not taking the time to prevent the
change.