Add acquired date field to default search result return fields
DISCOVERYACCESS-3884

This isn't generally needed for the search result, but we want to try to
keep the rss feed (which does require it) on the same Solr request
handler as search for as long as its needs don't diverge significantly.
The acquired_dt is a single timestamp present on 8% to 9% of records, so
it is small enough not to significantly impact default search behavior
even if it isn't needed most of the time. If the rss service grows to
need a lot more data, we may want to revisit the possibility of moving
it to a separate Solr request handler.