Expand list of tokens we don't strip periods off for facets
We've always stripped facet values of their trailing periods because the
presence of ISBD periods is irregular. When the headings end with
periods that are part of the data rather than part of the ISBD standard,
this can end up looking awkward (e.g. "Smith, A.B"). We've previously
re-added the periods for headings ending with a single initial or with
"etc", but this new update expands this to use a static list of tokens
that should be followed by non-ISBD periods. (Currently, "etc", "Jr",
"Sr", "Inc", "Co".) This list can be added to. It can't easily be made
language-dependent, as we can't easily determine the language of our
metadata values.