Update tests
Some out of date test files are removed entirely. A test in HoldingsAndItemsTest.java that now fails because it uses live catalog data is now disabled. This functionality will soon be removed, so repairing the test seemed unnecessary. Finally, a couple of tests were updated because they too depend on live data - in this case authority records. This dependency should be better encapsulated for testing in the future, but is fairly stable for now due to the slow rate of authority record changes.