Set heading work count instead of incrementing
I originally pushed the counts to the database as "value = value + x" instead of "value = x" because I had multiple facets that would add to the same stats. This proved to be problematic, and now no two facets values will ever need to be summed for a browse heading count stat.