Drop CurrentDBTable enum
Using this table to keep track of database table names in the Solr inventory database has not done any good w/r/t improved tracking of the tables and where they are used. Direct use of the table names works just as well as all the names are unique enough for effective grepping. It also has the benefit of preventing the SpotBugs app from flagging all of the queries that use the tables as SQL-injection risks due to their concatenating disparate strings to form queries.