Don't load subfield into MarcRecord if no subfield code
In Marc21, the first character in the subfield is automatically the subfield code, making it impossible to encode a subfield with no code that has any content at all. So dropping the entire subfield is risk free. In MARC XML or Json, this is no longer a limitation of the format, making it a potential risk once the data is stored in one of these formats.