Move HeadTypeDesc/HeadType to HeadingType/HeadingCategory
The HeadType enum, now HeadingCategory, describes the ways in which headings may be used. e.g authors or subjects.

The HeadTypeDesc enum, now HeadingType, describes the kinds of headings that might be authorized. e.g. persons or places.

The intersection of these two is described by the enum BlacklightField, e.g. a SUBJECT_PERSON describes a heading of HeadingType.PERSNAME used in a bibliographic field of HeadingCategory.SUBJECT.