Re-add parent to work heading entry
Indexing based on headings split between parent (author) and child (title) turned out to be a clumsy way to push things into a production indexing model. With this update, I'm trying to keep the parent-child links for how they may come into play for DISCOVERYACCESS-3069, but with the complete heading attached to the child (work) heading entry. When we get into bibliographic record headings and their higher level of encoding errors, the code will need to be able to work with the inevitable if low frequency error where a work entry it placed in the wrong field and creates a conflict with the same heading in the correct field. (e.g. a 710 Shakespeare's Romeo & Juliet will have a corporate name parent entry of William Shakespeare, while many more similar fields will be correctly in 700 fields, linking to William Shakespeare as a person.)

DISCOVERYACCESS-4172