Use first 008 date for sort when would otherwise use missing 2nd date
DISCOVERYACCESS-4870

The date type p indicates that the two 008 dates should be distribution date and production date respectively, while for date type r they should be reprint date and original date. In both cases, we choose to use the second date as the primary sort date. But, when the 2nd date is missing, we'll default back to the 1st. When either date is non-specific (e.g. 17uu), we'll use the equivalent zero-ed specific value (e.g. 1700), but consider it a valid date in the context of sort. (1uuu is too vague, and will not be used under this implementation.)