Exclude non-TOU 856$i values from TOU field extraction
So far it looks like mostly GPO links that are including both the $i TOU data, identifying dbcode, providercode, and sometimes a $i field like "Archived issues" or "2017 Issue" that seems like it might better belong in a $z. Though these are likely intended for display, I'm not sure how dependable that judgement would be, so I'm just making sure that they don't interfere with the TOU field extraction for now.

DISCOVERYACCESS-5149