Have AuthorTitle.java work with a copy of 245 fields
AuthorTitle.java makes changes to the 245$h before concatenating several different title fields. Because the field is shared with other generator modules, their results can be impacted depending on the order the modules are run, resulting in non-deterministic results in MARC.java and Format.java. It's easier to make a copy of the fields for AT.java to work with than to undertake the restructuring that would be necessary.

DISCOVERYACCESS-3071