Update project Gradle version
We've had a really old Gradle version specified in the project for quite a while. Eclipse is no longer accepting of Gradle versions less than 2.9. The current version is 5.x, but isn't compatible with our current Gradle config. At some point it'll be a good idea to get into the details of what feature(s) it includes that are incompatible, but updating to the most recent <5 release should be enough for now.