Attempt to address a minor looping glitch
I'm not sure whether this will help, but execution is sometimes moving from inside the while loop at line 78 and returning to top of the the while loop at line 55. I'm not sure why this is happening, but I'm eliminating the date comparison at line 88 as a possible, if unlikely culprit.