Don't suppress older acquisition date facet population
Because the integration tests are run against real, live data, the two year limit on the acquisition date fields' population means that the tests will need to be updated every couple of years to use new examples. This means a notable simplification of the logic in NewBooks.java, but the changes aren't as extensive as the diff suggests. I found that the indentation was irregular, and touched most of the file while cleaning it up.