Restructure acquisition date validation to report all invalid dates
I had been logging the acquisition dates that were eight digits, as expected, but weren't valid dates in the form of yyyymmdd. The restructured code will log all the invalid dates, as well as create a flag in the solr index for invalid dates, making it easier to report on bad dates as logged invalid dates may have been fixed since being logged.