Block suppressed headings counts from browse
This affects author, subject, and author-title browse. The "search" query handler is already being automatically filtered, but switching these queries from the "standard" handler to the "search" handler would require a good number of other changes. Not ideal when this entire job is deprecated once iterative browse updates is in production.
DISCOVERYACCESS-5823