Identify batch item updates using update volume
This is an imperfect test, but because the process checks every few moments for item changes, it will quickly revert to processing changes as manual once a batch load finishes. There is a good chance that manual updates that occur during an item batch update will be mistaken for part of the batch process. These changes may be held up until the batch load lag clears, though if the person doing the editing proceeds to update a bib or holding record, that change should not be mistaken as a batch update and will push ahead any other changes that might be queued with lower priority.

DISCOVERYACCESS-5481