Improve relator display
A couple of minor improvements...

Currently the relators that get populated as LOC URI's are unrecognized and produce a logged error every time. Instead, we should use the data as data, because this is growing in popularity.

Also, get rid of some extraneous periods before relator terms. These are cataloged with the extra periods when the relators are encoded as $4 relator codes rather than $e relator terms, but they shouldn't be there when we display them as terms.