Create template for modeling code with LMS-related code configurable
This now works in the entire test suite. All of the production uses of DownloadMARC expect to have Voyager identified in the config file. The rest of the code in the project that connects to Voyager needs the same added level of indirection, so the Folio equivalents can be added without breaking the Voyager code. Some interactions with Voyager will either not have equivalents in Folio or Folio will have better options, so that will need to be dealt with over time.

DISCOVERYACCESS-5482