Create discogs_display field for discogs identifiers
DISCOVERYACCESS-6490

(I am deliberately not incrementing the version of this code. There are no discogs displayed populated in Voyager so far, so there's no need to trigger rerunning of this code against records that haven't changed. Once in production, the new code will take effect for any records that do change.)