Adjust quotations in <?xml> tag
The new Java library we're using applies single quotes in the <?xml> tag, while the old ones all used double. Rather than spending a lot of time figuring out which versions use which option, it's enough to make sure the output is consistent to avoid false positives in the change detection.