Add checks in case of corrupted SRS record
This should treat the corrupt record as though it doesn't exist and index as an instance-only bib. This isn't fully robust, but this error has only happened once so far. I can make it more thorough in the future if we get different types of corruption. In this case, the parsedRecord block was missing and we got an NPE when trying to pull data from it.