Guard against invalid syntax in holdings URI data
There were only a couple examples of this in the holdings bookplate links, but there may be more over time as more are created in Folio, as those will be more prone to error than the ones migrated from Voyager.
DISCOVERYACCESS-7492

The actual bad links are transplanted into the new test method.