Initial work on detecting changed Hathi Bibs
The code to check the Folio MARC 005 date to decipher Voyager mod dates for records not updated in Folio since our migration turned out to be unnecessary, because batch jobs have touched all the records in Folio. (But I haven't stripped it yet.) Instead, the focus now on comparing the details of the local record with those in Zephyr and HathiTrust itself to determine whether we have substantive changes to push. This iteration pulls all three versions but doesn't yet know how to do the comparison.

DISCOVERYACCESS-7591