Tweak UI validation
Java's built in URI() contructor does validation, but it appears to have been both too strict in some places and not strict enough in others.

It is flagging links as invalid if they contain spaces, double quotes, or curly brackets, but the links that contains those will still work in a browser as the browser will apply the necessary encodings. As a result, I'm now applying some very careful URL encoding to those characters in an effort not to hide other more problematic issues from the validation.

It also fails to flag values like "NOT A LINK" as invalid links (once the spaces are encoded and not a cause for rejection themselves). I suspect these are accepted because they may potentially be relative links, but we shouldn't have relative links in Folio. As a result, I'm now rejecting any links that don't contain a host component.