Use non-880 subject field's 2nd indicator to determine source vocab
When an 880 field and a non-880 field are linked as non-Roman and Romanized equivalents, we generally give the 800 (non-Roman) value precedence, as that is the original value even when the derivative non-880 (Romanized) is recognized by name authority.

When the non-880 (Romanized) value is authorized, both fields should have a 2nd indicator reflecting that, but it's a common error to flag the 880 (non-Roman) value as unauthorized because the non-Roman value IN THAT FIELD, is not specifically authorized.

So, given the higher reliability of the 2nd indicator in the Romanized field, we can simply give precedence to that indicator, making this a very simple fix.

DISCOVERYACCESS-8135