Stop flipping already closed date range when one of the years slides
FOL-173
We were flipping the heading when one of the years changes by less than five years and the other stays the same, but it was decided this might be too risky to automate. We will continue to flip changes to the $d when both years remain the same but other details are updated.