Remove backslashes from search string everywhere
Backslash isn't part of a valid heading, and can potentially cause search failures.

org.apache.solr.search.SyntaxError: Non-hex character in Unicode escape sequence: k

error was returned when searching for
 \U+00d0\ukić, Bogomir, 1943-

Interestingly, Solr did not object to  \U+00d0 which is the encoding for capital eth (Đ), and part of the intended name, "Đukić". Removing the backslash will invalidate the Unicode escape character sequence for Solr, but that is already an incorrect encoding in this context. A search for this string without the backslashes will not find anything, but it's outside scope to attempt to correct corrupted values.