Adjust filter flags for hathi links to differentiate from regular links
The filter flags url_access and url_other were intended to differentiate between links created from Folio links from supplementary HathiTrust links added based on hathifiles data. But because the HathiLinks.java logic uses the URL.java work, the flags were populated in both places.

Additionally, the test suite needed to be updated with the change. I also uncommented a test for an in-copyright link, and one for a link pulled in using an OCLC number, as both major cases were missing.

DACCESS-591