Expand block on suggesting flips to current form
I previously blocked suggested changes if the most recent form of a heading matches either a 1xx older than the heading being evaluated (i.e. the heading reverted later to an older form), or a 4xx on the record being evaluated or an older version of the same record (i.e. what was once an alternate form will later become the preferred form).

This small tweak expands that to cases where the most recent version of a heading is a variant varying only in diacritics. capitalization, and/or punctuation from the 1xx of the currently evaluated record.
FOL-254